app/requirements.txt
Flask==3.0.0
openai==1.6.1
pymongo==4.6.1
python_bcrypt==0.3.2
requests==2.26.0
python-json-logger==2.0.7
